+++ tests/macros/test_auto.py
@@ -8,12 +8,12 @@
 
 from __future__ import annotations
 
-import asyncio
-
 from typing import TYPE_CHECKING
 
 import pytest
 from pytest_mock import MockerFixture
+
+from clu.command import Command
 
 from hal import config
 from hal.macros.auto import AutoModeMacro
@@ -30,7 +30,13 @@ def mock_auto_macro(mocker: MockerFixture, actor: HALActor):
     macro = actor.helpers.macros["auto"]
     orig_run = macro.run
 
-    macro.run = mocker.AsyncMock()
+    async def cancel_macro():
+        # Cancel macro after one iteration.
+        macro.cancelled = True
+        return True
+
+    macro.run = mocker.AsyncMock(side_effect=cancel_macro)
+    macro.command = Command(actor=actor)
 
     yield actor.helpers.macros["auto"]
 
@@ -42,18 +48,23 @@ def test_auto_macro(actor: HALActor):
     assert isinstance(macro, AutoModeMacro)
 
 
-async def test_auto_command(actor: HALActor, mock_auto_macro: AutoModeMacro):
-    async def stop_loop():
-        await asyncio.sleep(0.2)
-        mock_auto_macro.cancelled = True
-
-    asyncio.create_task(stop_loop())
+async def test_auto_command(
+    actor: HALActor,
+    mock_auto_macro: AutoModeMacro,
+    mocker: MockerFixture,
+):
+    mock_auto_macro.reset = mocker.AsyncMock()
 
     cmd = actor.invoke_mock_command("auto")
     await cmd
 
     assert cmd.status.did_succeed
+
     mock_auto_macro.run.assert_called()
+    mock_auto_macro.reset.assert_called_with(cmd, count=1, preload_ahead_time=None)
+
+    assert mock_auto_macro.config["count"] == 1
+    assert mock_auto_macro.config["preload_ahead_time"] == 300
 
 
 @pytest.mark.parametrize(
@@ -94,3 +105,16 @@ async def test_auto_expose_time(
         wait_time,
         config["macros"]["auto"]["preload_ahead_time"],
     )
+
+
+async def test_auto_macro_preload_ahead(
+    actor: HALActor,
+    mock_auto_macro: AutoModeMacro,
+    mocker: MockerFixture,
+):
+    mock_auto_macro.reset = mocker.AsyncMock()
+    cmd = actor.invoke_mock_command("auto --preload-ahead 100")
+    await cmd
+
+    assert cmd.status.did_succeed
+    mock_auto_macro.reset.assert_called_with(cmd, count=1, preload_ahead_time=100)